+++ backend/src/main/java/com/cherish/backend/config/SecurityConfig.java
@@ -1,51 +1,26 @@
 package com.cherish.backend.config;
 
-import com.cherish.backend.controller.argumentresolver.AvatarArgumentResolver;
-import com.cherish.backend.infra.interceptor.LoginCheckInterceptor;
-import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import java.util.List;
-
-@RequiredArgsConstructor
 @Configuration
-public class SecurityConfig implements WebMvcConfigurer{
-
-    private final AvatarArgumentResolver argumentResolver;
-    private final LoginCheckInterceptor loginCheckInterceptor;
+public class SecurityConfig {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
 
         http
-                .headers(config -> config.frameOptions( f -> f.disable()));
+                .headers(config -> config.frameOptions(f -> f.disable()));
 
         http
                 .authorizeHttpRequests(config -> config.anyRequest().permitAll())
                 .csrf(AbstractHttpConfigurer::disable);
 
         return http.build();
-    }
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        WebMvcConfigurer.super.addArgumentResolvers(resolvers);
-        resolvers.add(argumentResolver);
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(loginCheckInterceptor)
-                .addPathPatterns("/**")
-                .excludePathPatterns("/api/account/oauthlogin", "/api/account/tokenlogin", "/api/account/anotherplatformlogin", "/api/account/signup")
-                .excludePathPatterns("/", "/error");
     }
 
 }
