+++ backend/src/main/java/com/cherish/backend/config/WebConfig.java
@@ -1,0 +1,34 @@
+package com.cherish.backend.config;
+
+import com.cherish.backend.controller.argumentresolver.AvatarArgumentResolver;
+import com.cherish.backend.infra.interceptor.LoginCheckInterceptor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    private final AvatarArgumentResolver argumentResolver;
+    private final LoginCheckInterceptor loginCheckInterceptor;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        WebMvcConfigurer.super.addArgumentResolvers(resolvers);
+        resolvers.add(argumentResolver);
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(loginCheckInterceptor)
+                .addPathPatterns("/**")
+                .excludePathPatterns("/api/test/destroy")
+                .excludePathPatterns("/api/account/oauthlogin", "/api/account/tokenlogin", "/api/account/anotherplatformlogin", "/api/account/signup")
+                .excludePathPatterns("/", "/error","/h2-console","/favicon.ico","/error");
+    }
+}
